split grpc and server code
This commit splits gRPC API away from server code
that handles SPDK startup and monitoring.
